+++ salt/_pillar/metalk8s.py
@@ -129,6 +129,7 @@ def _load_networks(config_data):
         "pod": networks_data.get("pods", DEFAULT_POD_NETWORK),
         "service": networks_data.get("services", DEFAULT_SERVICE_NETWORK),
         "portmap": networks_data.get("portmap"),
+        "nodeport": networks_data.get("nodeport"),
     }