+++ salt/tests/unit/formulas/fixtures/salt.py
@@ -457,6 +457,9 @@ register_basic("metalk8s_network.get_control_plane_ingress_endpoint")(
 register_basic("metalk8s_network.get_portmap_ips")(
     MagicMock(return_value=["192.168.1.100", "127.0.0.1"])
 )
+register_basic("metalk8s_network.get_nodeport_cidrs")(
+    MagicMock(return_value=["192.168.2.0/24", "127.0.0.1/32"])
+)
 register_basic("metalk8s_network.get_control_plane_ingress_external_ips")(
     MagicMock(
         return_value=[