+++ CHANGELOG.md
@@ -46,8 +46,8 @@
 ### Enhancements
 
 - Bump Kubernetes version to
-  [1.23.5](https://github.com/kubernetes/kubernetes/releases/tag/v1.23.5)
-  (PR[#3727](https://github.com/scality/metalk8s/pull/3727))
+  [1.23.8](https://github.com/kubernetes/kubernetes/releases/tag/v1.23.8)
+  (PR[#3812](https://github.com/scality/metalk8s/pull/3812))
 
 - Bump etcd version to [3.5.1](https://github.com/etcd-io/etcd/releases/tag/v3.5.1)
   (PR[#3634](https://github.com/scality/metalk8s/pull/3634))